requirements.txt
asgiref==3.9.2
Django==5.2.7
django-phonenumber-field==8.3.0
phonenumbers==9.0.16
phonenumberslite==9.0.16
pillow==11.3.0
sqlparse==0.5.3
